Add security, ,authentication, jwtToken

+++ src/main/java/com/example/eSmartRecruit/repositories/ApplicationRepos.java
@@ -1,11 +1,10 @@
 package com.example.eSmartRecruit.repositories;
 
+import com.example.eSmartRecruit.models.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.eSmartRecruit.models.Application;
-
 @Repository
-public interface ApplicationRepos extends JpaRepository<Application, Long>{
+public interface ApplicationRepos extends JpaRepository<Application, Integer>{
 
 }
