+++ src/main/java/com/example/eSmartRecruit/repositories/InterviewSessionRepos.java
@@ -1,11 +1,10 @@
 package com.example.eSmartRecruit.repositories;
 
+import com.example.eSmartRecruit.models.InterviewSession;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.eSmartRecruit.models.InterviewSession;
-
 @Repository
-public interface InterviewSessionRepos extends JpaRepository<InterviewSession, Long>{
+public interface InterviewSessionRepos extends JpaRepository<InterviewSession, Integer>{
 
 }
