+++ src/main/java/com/example/eSmartRecruit/repositories/ReportRepos.java
@@ -1,11 +1,10 @@
 package com.example.eSmartRecruit.repositories;
 
+import com.example.eSmartRecruit.models.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.eSmartRecruit.models.Report;
-
 @Repository
-public interface ReportRepos extends JpaRepository<Report, Long>{
+public interface ReportRepos extends JpaRepository<Report, Integer>{
 
 }
