+++ src/main/java/com/example/eSmartRecruit/repositories/SkillRepos.java
@@ -1,11 +1,10 @@
 package com.example.eSmartRecruit.repositories;
 
+import com.example.eSmartRecruit.models.Skill;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.eSmartRecruit.models.Skill;
-
 @Repository
-public interface SkillRepos extends JpaRepository<Skill, Long>{
+public interface SkillRepos extends JpaRepository<Skill, Integer>{
 
 }
